+++ test/test_StreamOps.cpp
@@ -19,7 +19,7 @@
 #include "doctest.h"
 #include "ArrHelpers.hpp"
 #include "ZArr.hpp"
-#include "Testability.hpp"
+#include "StreamOps.hpp"
 
 // non-vectorized version for comparison
 void hann_calc(Z* out, int n) {
